User commit

+++ src/main/java/com/example/demo/repository/StudentRepo.java
@@ -3,6 +3,6 @@ import org.springframework.data,jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 import com.example.demo.entity.Student;
 @Repository
-public interface StudentRepo extends JpaRepository<Student,Long>{
+public interface StudentRepo extends JpaRepository<Student,Integer>{
     
 }
